break: split data from code to reduce bundle sizes
Bundlers often include the full data set which increases final bundle
sizes unnecessarily.
This breaking change separates code from data.

Please use:

import astronomia from 'astronomia' // for code
import data from 'astronomia/data' // for data
