+++ src/index.js
@@ -6,7 +6,6 @@ export { default as binary } from './binary.js'
 export { default as coord } from './coord.js'
 export { default as conjunction } from './conjunction.js'
 export { default as circle } from './circle.js'
-export { default as data } from '../data/index.js'
 export { default as deltat } from './deltat.js'
 export { default as elementequinox } from './elementequinox.js'
 export { default as elliptic } from './elliptic.js'
