+++ test/jupitermoons.test.js
@@ -1,6 +1,7 @@
 import assert from 'assert'
 import float from './support/float.js'
-import { julian, jupitermoons, data, planetposition } from '../src/index.js'
+import { julian, jupitermoons, planetposition } from '../src/index.js'
+import data from '../data/index.js'
 
 describe('#jupitermoons', function () {
   describe('positions', function () {
