+++ test/perihelion.test.js
@@ -2,7 +2,8 @@
 
 import assert from 'assert'
 import float from './support/float.js'
-import { julian, perihelion, data, planetposition } from '../src/index.js'
+import { julian, perihelion, planetposition } from '../src/index.js'
+import data from '../data/index.js'
 
 function toObj (props, table) {
   return table.map(function (row) {
