+++ test/pluto.test.js
@@ -1,6 +1,7 @@
 import assert from 'assert'
 import float from './support/float.js'
-import { planetposition, data, pluto, sexagesimal as sexa } from '../src/index.js'
+import { planetposition, pluto, sexagesimal as sexa } from '../src/index.js'
+import data from '../data/index.js'
 
 describe('#pluto', function () {
   it('heliocentric()', function () {
