+++ test/saturnmoons.test.js
@@ -2,7 +2,8 @@
 
 import assert from 'assert'
 import float from './support/float.js'
-import { saturnmoons, planetposition, data } from '../src/index.js'
+import { saturnmoons, planetposition } from '../src/index.js'
+import data from '../data/index.js'
 
 describe('#saturnmoons', function () {
   function comp (res, exp) {
